LayoutTests/crypto/workers/subtle/raw-postMessage-worker-expected.txt

Test sending raw crypto keys via postMessage to a worker.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS All checks passed in worker
PASS key.type is 'secret'
PASS key.extractable is false
PASS key.algorithm.name is 'PBKDF2'
PASS key.usages is ['deriveBits', 'deriveKey']
PASS successfullyParsed is true

TEST COMPLETE
